Added class Player and GameLayout module
